+++ CHANGELOG.md
@@ -78,6 +78,11 @@
   * Added `asset_id` schema field to `akamai_property_bootstrap`.
 
 
+
+* PAPI
+  * Added validation to raise an error when updating `contact` and `auto_acknowledge_rule_warnings` fields on active property versions in `akamai_property_activation` resource.
+
+
 #### BUG FIXES:
 
 
@@ -135,7 +140,6 @@
 * PAPI
   * Fixed improper diff in `akamai_property` resource when no value was given in the child for
     the field `criteria_must_satisfy`, which the API sets to `all` by default.
-